@@ -1,6 +1,6 @@
 {
     "name": "@andes/app",
-    "version": "2.6.2",
+    "version": "3.0.0",
     "description": "Aplicación web Angular2 para ANDES",
     "angular-cli": {},
     "scripts": {
@@ -9,8 +9,8 @@
         "test": "ng test",
         "lint": "ng lint",
         "e2e": "ng e2e",
-        "build:prod": "node --max_old_space_size=8192 ./node_modules/@angular/cli/bin/ng build --env=prod --prod --aot=true",
-        "build:test": "node --max_old_space_size=8192 ./node_modules/@angular/cli/bin/ng build --env=test"
+        "build:prod": "ng build --env=prod --prod --aot=true",
+        "build:test": "ng build --env=test"
     },
     "repository": {
         "type": "git",
@@ -22,44 +22,44 @@
         "url": "https://github.com/andes/app/issues"
     },
     "dependencies": {
-        "@andes/auth": "^0.3.0",
+        "@agm/core": "^1.0.0-beta.2",
+        "@andes/auth": "^2.0.0",
         "@andes/match": "^0.1.9",
-        "@andes/plex": "^2.0.14",
-        "@andes/shared": "^1.0.1",
-        "@angular/animations": "^4.3.1",
-        "@angular/cdk": "2.0.0-beta.8",
-        "@angular/common": "^4.3.1",
-        "@angular/compiler": "^4.3.1",
-        "@angular/core": "^4.3.1",
-        "@angular/forms": "^4.3.1",
-        "@angular/http": "^4.3.1",
-        "@angular/material": "2.0.0-beta.8",
-        "@angular/platform-browser": "^4.3.1",
-        "@angular/platform-browser-dynamic": "^4.3.1",
-        "@angular/platform-server": "^4.3.1",
-        "@angular/router": "^4.3.1",
-        "angular2-google-maps": "^0.17.0",
+        "@andes/plex": "^3.1.0",
+        "@andes/shared": "^2.0.0",
+        "@angular/animations": "^5.0.0",
+        "@angular/cdk": "^5.2.4",
+        "@angular/common": "^5.0.0",
+        "@angular/compiler": "^5.0.0",
+        "@angular/core": "^5.0.0",
+        "@angular/forms": "^5.0.0",
+        "@angular/http": "^5.0.0",
+        "@angular/material": "^5.2.4",
+        "@angular/platform-browser": "^5.0.0",
+        "@angular/platform-browser-dynamic": "^5.0.0",
+        "@angular/platform-server": "^5.0.0",
+        "@angular/router": "^5.0.0",
         "bowser": "^1.7.1",
         "bson": "^1.0.4",
         "chart.js": "^2.7.1",
         "core-js": "^2.4.1",
         "enhanced-resolve": "^3.4.1",
-        "file-saver": "^1.3.3",
+        "file-saver": "^1.3.8",
         "ng2-charts": "^1.6.0",
         "ng2-drag-drop": "^1.5.1",
         "ng2-slugify": "^0.1.0",
         "rxjs": "^5.4.2",
-        "zone.js": "^0.8.14"
+        "zone.js": "^0.8.20"
     },
     "devDependencies": {
-        "@angular/cli": "~1.3.2",
-        "@angular/compiler-cli": "^4.3.1",
+        "@angular/cli": "^1.7.3",
+        "@angular/compiler-cli": "^5.0.0",
         "@types/jasmine": "^2.2.30",
-        "@types/node": "^6.0.42",
+        "@types/node": "^9.4.7",
         "codelyzer": "~2.0.0-beta.4",
         "jasmine-core": "~2.5.2",
         "jasmine-spec-reporter": "~3.2.0",
-        "karma": "~1.4.1",
+        "karma": "^1.7.1",
         "karma-chrome-launcher": "~2.0.0",
         "karma-cli": "~1.0.1",
         "karma-coverage-istanbul-reporter": "^0.2.0",
@@ -68,7 +68,6 @@
         "protractor": "~5.1.0",
         "ts-node": "~2.0.0",
         "tslint": "~4.4.2",
-        "typescript": "~2.4.2",
-        "webpack": "^3.11.0"
+        "typescript": "~2.6.2"
     }
 }
