@@ -18,21 +18,22 @@ import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
 import { LOCALE_ID } from '@angular/core';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
+import { HttpClientModule } from '@angular/common/http';
 import { HttpModule } from '@angular/http';
 
 // Global
 import { PlexModule } from '@andes/plex';
 import { Plex } from '@andes/plex';
 import { Server } from '@andes/shared';
+import { AuthModule } from '@andes/auth';
 import { Auth } from '@andes/auth';
 import { RoutingGuard } from './app.routings-guard.class';
-import { AgmCoreModule } from 'angular2-google-maps/core';
+import { AgmCoreModule } from '@agm/core';
 import { MapsComponent } from './utils/mapsComponent';
 import { PermisosComponent } from './utils/permisos/permisos.component';
 import { Ng2DragDropModule } from 'ng2-drag-drop';
 import { HoverClassDirective } from './directives/hover-class.directive';
 import { DocumentosService } from './services/documentos.service';
-
 
 // Pipes
 import { EdadPipe } from './pipes/edad.pipe';
@@ -226,7 +227,9 @@ import { EjecucionInternacionComponent } from './modules/rup/components/ejecucio
 import { EgresoInternacionComponent } from './modules/rup/components/elementos/egresoInternacion.component';
 import { OcuparCamaComponent } from './modules/rup/components/ejecucion/internacion/ocuparCama.component';
 import { PasesCamaComponent } from './modules/rup/components/elementos/pasesCama.component';
-
+import { InformeEpicrisisComponent } from './modules/rup/components/elementos/informeEpicrisis.component';
+import { OdontologiaDefaultComponent } from './modules/rup/components/elementos/odontologiaDefault.component';
+import { CircunferenciaCinturaComponent } from './modules/rup/components/elementos/circunferenciaCintura.component';
 
 // TODO: Eliminar todo esto de las llaves: deprecated
 import { LlavesTipoPrestacionComponent } from './components/llaves/tipoPrestacion/llaves-tipoPrestacion.component';
@@ -250,6 +253,10 @@ import { ArbolPermisosComponent } from './components/usuario/arbolPermisos.compo
 
 // REPORTES
 import { ReporteC2Component } from './components/reportes/reporteC2.component';
+import { ConsultaDiagnosticoComponent } from './components/reportes/consultaDiagnostico.component';
+import { CantidadConsultaXPrestacionComponent } from './components/reportes/cantidadConsultaXPrestacion.component';
+import { EncabezadoReportesComponent } from './components/reportes/encabezadoReportes.component';
+
 
 // Locales
 import { AppComponent } from './app.component';
@@ -310,8 +317,11 @@ export let RUPRegistry = {
     'IngresoInternacionComponent': IngresoInternacionComponent,
     'EgresoInternacionComponent': EgresoInternacionComponent,
     'PasesCamaComponent': PasesCamaComponent,
+    'InformeEpicrisisComponent': InformeEpicrisisComponent,
     'OtoemisionAcusticaDeOidoDerechoComponent': OtoemisionAcusticaDeOidoDerechoComponent,
     'OtoemisionAcusticaDeOidoIzquierdoComponent': OtoemisionAcusticaDeOidoIzquierdoComponent,
+    'OdontologiaDefaultComponent': OdontologiaDefaultComponent,
+    'CircunferenciaCinturaComponent': CircunferenciaCinturaComponent,
 };
 
 let RUPComponentsArray = [
@@ -346,11 +356,20 @@ let RUPComponentsArray = [
     IniciarInternacionComponent,
     EjecucionInternacionComponent,
     EgresoInternacionComponent,
-    PasesCamaComponent
+    PasesCamaComponent,
+    InformeEpicrisisComponent,
+    OdontologiaDefaultComponent,
+    CircunferenciaCinturaComponent
 ];
-// for (let key in RUPRegistry) {
-//     RUPComponentsArray.push(RUPRegistry[key]);
-// }
+
+import { EstadisticaModule } from './modules/estadisticas/estadistica.module';
+
+/** moment pipes  - desde agular 5 hay que importar el locale a demanda */
+import { registerLocaleData } from '@angular/common';
+import localeEs from '@angular/common/locales/es';
+
+registerLocaleData(localeEs, 'es');
+
 
 // Main module
 @NgModule({
@@ -358,11 +377,14 @@ let RUPComponentsArray = [
         BrowserModule,
         ReactiveFormsModule,
         FormsModule,
+        HttpClientModule,
         HttpModule,
         PlexModule,
+        AuthModule,
         Ng2DragDropModule,
         ChartsModule,
         routing,
+        EstadisticaModule,
         AgmCoreModule.forRoot({
             apiKey: 'AIzaSyAJuFVuMmVwV8gtP_1m3Ll1VzHagAI_X9I'
         })
@@ -399,6 +421,9 @@ let RUPComponentsArray = [
         SolicitudTurnoVentanillaComponent, ListaSolicitudTurnoVentanillaComponent, ActivarAppComponent,
         BusquedaUsuarioComponent, UsuarioCreateComponent, UsuarioUpdateComponent,
         ReporteC2Component,
+        ConsultaDiagnosticoComponent,
+        CantidadConsultaXPrestacionComponent,
+        EncabezadoReportesComponent,
         ListarTurnosComponent, ListarCarpetasComponent,
         MapaEspacioFisicoComponent, SuspenderAgendaComponent,
         ResumenComponent,
@@ -439,6 +464,7 @@ let RUPComponentsArray = [
             useValue: 'es-AR'
         },
         Plex,
+        Server,
         Auth,
         RoutingGuard,
         OrganizacionService,
@@ -461,7 +487,6 @@ let RUPComponentsArray = [
         TurnoService,
         EspacioFisicoService,
         ListaEsperaService,
-        Server,
         SmsService,
         PrestacionesService,
         AdjuntosService,
