@@ -170,6 +170,8 @@ export class ListarSolicitudesComponent implements OnInit {
                 listaProfesionales = resultado;
                 event.callback(listaProfesionales);
             });
+        } else {
+            event.callback(this.espacioFisico || []);
         }
     }
 
@@ -191,7 +193,10 @@ export class ListarSolicitudesComponent implements OnInit {
 
     switchMarcarTodas() {
         this.marcarTodas = !this.marcarTodas;
-        this.carpetasSeleccionadas = this.marcarTodas ? this.carpetas : [];
+        this.carpetasSeleccionadas = this.marcarTodas ? this.carpetas.filter(function (el) {
+            return el.estado === 'En archivo';
+        }) : [];
+        this.verPrestar = false;
     }
 
     prestarCarpetas() {
@@ -227,15 +232,23 @@ export class ListarSolicitudesComponent implements OnInit {
     }
 
     prestar(solicitudCarpeta) {
-        this.carpetaPrestadaEmit.emit(solicitudCarpeta);
-        this.carpetaSeleccionada = solicitudCarpeta;
-        this.verPrestar = true;
+        if (this.carpetasSeleccionadas.length === 0) {
+            this.carpetaPrestadaEmit.emit(solicitudCarpeta);
+            this.carpetaSeleccionada = solicitudCarpeta;
+            this.verPrestar = true;
+        }
     }
 
-    sortCarpetas() {
+    sortCarpetas() { // se divide this.carpetas en letras y en numeros para hacer el sort correspondiente
         let val = this.sortDescending ? -1 : 1;
-        // this.carpetas.sort((a, b) => { return (parseInt(a.numero) > parseInt(b.numero)) ? val : (parseInt(b.numero) > parseInt(a.numero)) ? -val : 0; } );
-        this.carpetas.sort((a, b) => { return (a.numero > b.numero) ? val : ((b.numero > a.numero) ? -val : 0); });
+        let carpetas_numeros = this.carpetas.filter(x => !isNaN(x.numero));
+        let carpetas_letras = this.carpetas.filter(x => isNaN(x.numero));
+        carpetas_letras.sort((a, b) => { return (a.numero > b.numero) ? val : (b.numero > a.numero) ? -val : 0; });
+        carpetas_numeros.sort((a, b) => { return (parseInt(a.numero, 10) > parseInt(b.numero, 10)) ? val : ((parseInt(b.numero, 10) > parseInt(a.numero, 10)) ? -val : 0); });
+
+        let carpetas_sort = carpetas_numeros.concat(carpetas_letras);
+        this.carpetas = [];
+        this.carpetas = carpetas_sort;
     }
 
     toogleSort() {
