@@ -25,6 +25,9 @@ export class PuntoInicioTurnosComponent implements OnInit {
     @Output() selected: EventEmitter<any> = new EventEmitter<any>();
     @Output() escaneado: EventEmitter<any> = new EventEmitter<any>();
 
+    public puedeCrearSolicitud = false;
+    public puedeAutocitar = false;
+    public puedeDarTurno = false;
     public alerta = false;
     public mostrarLista = true;
     public mostrarPacientesSearch = true;
@@ -58,7 +61,9 @@ export class PuntoInicioTurnosComponent implements OnInit {
         private plex: Plex) { }
 
     ngOnInit() {
-        this.autorizado = this.auth.getPermissions('turnos:darTurnos:?').length > 0;
+        this.autorizado = this.auth.getPermissions('turnos:puntoInicio:?').length > 0;
+        this.puedeDarTurno = this.auth.getPermissions('turnos:puntoInicio:darTurnos:?').length > 0;
+        this.puedeCrearSolicitud = this.auth.getPermissions('turnos:puntoInicio:solicitud:?').length > 0;
     }
 
     showArancelamientoForm(turno) {
