@@ -243,7 +243,7 @@ export class RevisionAgendaComponent implements OnInit {
             listaTurnos = listaTurnos.concat(this.agenda.sobreturnos);
         }
         turnoSinVerificar = listaTurnos.find(t => {
-            return (t && t.paciente && t.paciente.id && !t.asistencia && t.estado !== 'suspendido');
+            return (t && t.paciente && t.paciente.id && !t.asistencia && t.estado !== 'suspendido' && t.estado !== 'turnoDoble');
         });
         if (!turnoSinVerificar) { // Si todos los turnos están verificados..
             // Se cambia de estado la agenda a pendienteAuditoria
@@ -291,7 +291,6 @@ export class RevisionAgendaComponent implements OnInit {
                     && !t.diagnostico.ilegible && t.asistencia === 'asistio')) || !t.asistencia)
             );
         });
-
         if (!turnoSinCodificar) {
             // Se cambia de estado la agenda a asistenciaCerrada
             let patch = {
