@@ -17,6 +17,8 @@ import { OcupacionService } from '../../../../services/ocupacion/ocupacion.servi
 import { FinanciadorService } from '../../../../services/financiador.service';
 import { ProcedimientosQuirurgicosService } from '../../../../services/procedimientosQuirurgicos.service';
 import { Cie10Service } from '../../../../services/term/cie10.service';
+import { OrganizacionService } from '../../../../services/organizacion.service';
+
 @Component({
     selector: 'rup',
     styleUrls: [
@@ -47,10 +49,12 @@ export class RUPComponent implements OnInit {
      * @memberof RUPComponent
      */
     private loadComponent() {
+
         // Cargamos el componente
         const component = RUPRegistry[this.elementoRUP.componente];
         const componentFactory = this.componentFactoryResolver.resolveComponentFactory(component as any);
         const componentReference = this.viewContainerRef.createComponent(componentFactory);
+
         // Copia todas las propiedades
         componentReference.instance['prestacion'] = this.prestacion;
         componentReference.instance['registro'] = this.registro;
@@ -58,10 +62,12 @@ export class RUPComponent implements OnInit {
         componentReference.instance['soloValores'] = this.soloValores;
         componentReference.instance['paciente'] = this.paciente;
         componentReference.instance['params'] = this.params;
+
         // Event bubbling
         componentReference.instance['change'].subscribe(value => {
             this.emitChange(false);
         });
+
         // Inicia el detector de cambios
         componentReference.changeDetectorRef.detectChanges();
     }
@@ -81,7 +87,8 @@ export class RUPComponent implements OnInit {
         public sanitazer: DomSanitizer,
         public snomedService: SnomedService,
         public procedimientosQuirurgicosService: ProcedimientosQuirurgicosService,
-        public Cie10Service: Cie10Service
+        public Cie10Service: Cie10Service,
+        public servicioOrganizacion: OrganizacionService
     ) { }
 
     ngOnInit() {
