@@ -87,7 +87,6 @@ export class PrestacionEjecucionComponent implements OnInit {
     // el concepto que seleccionamos para eliminar lo guradamos aca.
     public conceptoAEliminar: any;
 
-
     public conceptosTurneables: any[];
 
     // Listado de grupos de la busqueda guiada
@@ -119,7 +118,6 @@ export class PrestacionEjecucionComponent implements OnInit {
         // Limpiar los valores observados al iniciar la ejecución
         // Evita que se autocompleten valores de una consulta anterior
         this.conceptObserverService.destroy();
-
         this.route.params.subscribe(params => {
             let id = params['id'];
             this.idAgenda = localStorage.getItem('idAgenda');
@@ -162,7 +160,6 @@ export class PrestacionEjecucionComponent implements OnInit {
                                     }
                                 }
                             }
-
                         }
                         this.elementosRUPService.guiada(this.prestacion.solicitud.tipoPrestacion.conceptId).subscribe((grupos) => {
                             this.grupos_guida = grupos;
@@ -657,6 +654,7 @@ export class PrestacionEjecucionComponent implements OnInit {
      * @memberof PrestacionEjecucionComponent
      */
     guardarPrestacion() {
+
         // validamos antes de guardar
         if (!this.beforeSave()) {
             return;
@@ -671,11 +669,13 @@ export class PrestacionEjecucionComponent implements OnInit {
 
         let params: any = {
             op: 'registros',
+            solicitud: this.prestacion.solicitud,
             registros: registros
         };
 
         this.servicioPrestacion.patch(this.prestacion.id, params).subscribe(prestacionEjecutada => {
-            this.plex.toast('success', 'Prestacion guardada correctamente', 'Prestacion guardada', 100);
+            this.plex.toast('success', 'Prestación guardada correctamente', 'Prestacion guardada', 100);
+
             // Si existe un turno y una agenda asociada, y existe un concepto que indica que el paciente no concurrió a la consulta...
             if (this.idAgenda) {
                 localStorage.removeItem('idAgenda');
@@ -703,10 +703,15 @@ export class PrestacionEjecucionComponent implements OnInit {
         });
     }
 
-    volver() {
-        this.plex.confirm('<i class="mdi mdi-alert"></i> Se van a perder los cambios no guardados', '¿Volver al Punto de Inicio?').then(confirmado => {
+    volver(ambito = 'ambulatorio') {
+        let mensaje = ambito === 'ambulatorio' ? 'Punto de Inicio' : 'Mapa de Camas';
+        this.plex.confirm('<i class="mdi mdi-alert"></i> Se van a perder los cambios no guardados', '¿Volver al ' + mensaje + '?').then(confirmado => {
             if (confirmado) {
-                this.router.navigate(['rup']);
+                if (ambito === 'ambulatorio') {
+                    this.router.navigate(['rup']);
+                } else {
+                    this.router.navigate(['mapa-de-camas']);
+                }
             } else {
                 return;
             }
