@@ -41,9 +41,9 @@ export class PrestacionesService {
     }
 
     /**
-     * Metodo get. Trae lista de objetos prestacion.
+     * Método get. Trae lista de objetos prestacion.
      *
-     * @param {*} params Opciones de busqueda
+     * @param {*} params Opciones de búsqueda
      * @param {*} [options={}] Options a pasar a la API
      * @returns {Observable<IPrestacion[]>}
      *
@@ -58,8 +58,9 @@ export class PrestacionesService {
 
         return this.server.get(this.prestacionesUrl, opt);
     }
+
     /**
-     * Metodo getById. Trae el objeto tipoPrestacion por su Id.
+     * Método getById. Trae el objeto tipoPrestacion por su Id.
      * @param {String} id Busca por Id
      */
     getById(id: String, options: any = {}): Observable<IPrestacion> {
@@ -72,7 +73,7 @@ export class PrestacionesService {
     }
 
     /**
-     * Metodo getByPaciente. Busca todas las prestaciones de un paciente
+     * Método getByPaciente. Busca todas las prestaciones de un paciente
      * @param {String} idPaciente
      */
     getByPaciente(idPaciente: any, recargarCache: boolean = false): Observable<any[]> {
@@ -99,7 +100,6 @@ export class PrestacionesService {
         }
 
     }
-
 
     findValues(obj, key) { // funcion para buscar una key y recupera un array con sus valores.
         return this.findValuesHelper(obj, key, []);
@@ -136,7 +136,7 @@ export class PrestacionesService {
 
 
     /**
-     * Metodo getByPacienteKey
+     * Método getByPacienteKey
      * @param {String} idPaciente
      */
     getByPacienteKey(idPaciente: any, key: any): Observable<any[]> {
@@ -159,7 +159,7 @@ export class PrestacionesService {
     }
 
     /**
-     * Metodo getByPacienteKey
+     * Método getByPacienteKey
      * @param {String} idPaciente
      */
     getRegistroById(idPaciente: any, id: any): Observable<any[]> {
@@ -178,7 +178,7 @@ export class PrestacionesService {
     }
 
     /**
-     * Metodo getByPacienteHallazgo lista todo los hallazgos registrados del paciente
+     * Método getByPacienteHallazgo lista todo los hallazgos registrados del paciente
      * @param {String} idPaciente
      */
     getByPacienteHallazgo(idPaciente: any, soloValidados?: boolean): Observable<any[]> {
@@ -336,7 +336,7 @@ export class PrestacionesService {
     }
 
     /**
-     * Metodo getByPacienteMedicamento lista todos los medicamentos registrados del paciente
+     * Método getByPacienteMedicamento lista todos los medicamentos registrados del paciente
      * @param {String} idPaciente
      */
     getByPacienteMedicamento(idPaciente: any, soloValidados?: boolean): Observable<any[]> {
@@ -422,7 +422,7 @@ export class PrestacionesService {
 
 
     /**
-     * Metodo getUnHallazgoPaciente x Concepto obtiene un hallazgo cronico o activo con todas sus evoluciones
+     * Método getUnHallazgoPaciente x Concepto obtiene un hallazgo cronico o activo con todas sus evoluciones
      * para un paciente
      * @param {String} idPaciente
      */
@@ -444,7 +444,7 @@ export class PrestacionesService {
     }
 
     /**
-     * Metodo getUnHallazgoPacienteXOrigen obtiene un hallazgo con todas sus evoluciones
+     * Método getUnHallazgoPacienteXOrigen obtiene un hallazgo con todas sus evoluciones
      * para un paciente buscandolo por el registro de origen
      * @param {String} idPaciente
      * @param {String} idRegistroOrigen
@@ -462,7 +462,7 @@ export class PrestacionesService {
 
 
     /**
-         * Metodo getUnMedicamentoXOrigen obtiene un registro de medicamento con todas sus evoluciones
+         * Método getUnMedicamentoXOrigen obtiene un registro de medicamento con todas sus evoluciones
          * para un paciente buscandolo por el registro de origen
          * @param {String} idPaciente
          * @param {String} idRegistroOrigen
@@ -480,7 +480,7 @@ export class PrestacionesService {
 
 
     /**
-     * Metodo getById. Trae el objeto tipoPrestacion por su Id.
+     * Método getById. Trae el objeto tipoPrestacion por su Id.
      * @param {String} id Busca por Id
      */
     getByKey(params: any, options: any = {}): Observable<IPrestacion[]> {
@@ -520,7 +520,7 @@ export class PrestacionesService {
     }
 
     /**
-     * Metodo post. Inserta un objeto nuevo.
+     * Método post. Inserta un objeto nuevo.
      * @param {any} prestacion Recibe solicitud RUP con paciente
      */
     post(prestacion: any): Observable<any> {
@@ -528,7 +528,7 @@ export class PrestacionesService {
     }
 
     /**
-     * Metodo put. Actualiza un objeto prestacionPaciente.
+     * Método put. Actualiza un objeto prestacionPaciente.
      * @param {IPrestacionPaciente} problema Recibe IPrestacionPaciente
      */
     put(prestacion: IPrestacion): Observable<IPrestacion> {
@@ -644,7 +644,9 @@ export class PrestacionesService {
     }
 
     validarPrestacion(prestacion, planes): Observable<any> {
+
         let planesCrear = undefined;
+
         if (planes.length) {
             planesCrear = [];
             planes.forEach(plan => {
@@ -671,6 +673,14 @@ export class PrestacionesService {
                         // asignamos la prestacion de origen
                         nuevaPrestacion.solicitud.prestacionOrigen = prestacion.id;
 
+                        if (plan.valor.solicitudPrestacion.organizacionDestino) {
+                            nuevaPrestacion.solicitud.organizacion = plan.valor.solicitudPrestacion.organizacionDestino;
+                        }
+
+                        if (plan.valor.solicitudPrestacion.profesionalesDestino) {
+                            nuevaPrestacion.solicitud.profesional = plan.valor.solicitudPrestacion.profesionalesDestino[0];
+                        }
+
                         // agregamos los registros en la solicitud
                         nuevaPrestacion.solicitud.registros.push(plan);
 
@@ -689,11 +699,9 @@ export class PrestacionesService {
         };
         return this.patch(prestacion.id, dto);
 
-
-
     }
     /**
-    * Metodo clonar. Inserta una copia de una prestacion.
+    * Método clonar. Inserta una copia de una prestacion.
     * @param {any} prestacionCopia Recibe una copia de una prestacion
     */
     clonar(prestacionCopia: any, estado: any): Observable<any> {
